bus: Make bus_driver_send_name_owner_changed take DBusConnections
All callers had access to the DBusConnection that is the old owner,
and the DBusConnection that is the new owner. Pass those to the
function instead of their names. This gives the function a bit more
contextual information, which will be important when we filter
NameOwnerChanged messages entering containers according to whether the
container is allowed to see the connection.

While I'm breaking its API anyway, rename the function to match the
signal that it emits, which was renamed in early 2005.

Signed-off-by: Simon McVittie <smcv@collabora.com>
Reviewed-by: Philip Withnall <withnall@endlessm.com>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=105656