bus: If containers block a requested reply, substitute an error reply
If some connection in a container has legitimately had one of its
methods called by another connection (either unconfined or in a
different container), but the reply is unsuitable for delivery to
the caller (perhaps because it contains fds), we don't want the
caller to be left waiting. Add a way for the security policy to
request that we send an error reply to the original method call, as
well as sending an error reply to the undesirable reply.

Signed-off-by: Simon McVittie <smcv@collabora.com>
Reviewed-by: Philip Withnall <withnall@endlessm.com>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=105656