Assert that can_see_well_known_name() isn't called for o.fd.DBus
The only call to bus_dispatch_matches() with
needs_to_see_well_known_name != NULL is in
bus_driver_send_name_owner_changed(), which is only called for bus
names tracked by the BusRegistry, so we want to assert that o.fd.DBus
can't be in the registry.

Similarly, assert that o.fd.DBus can't be in the activation registry,
which is true since commit "activation: Refuse to load activatable
services with invalid names".

Signed-off-by: Simon McVittie <smcv@collabora.com>
Reviewed-by: Philip Withnall <withnall@endlessm.com>