bus/connection: API for getting list of names from connections
The new function is going to be useful for checking policy
in non-linear way.

Now, each rule is matched against parameters,
which include a sender or a receiver. Thus, each name
appearing in a rule can be used in a query to the sender
or the receiver.

In non-linear way, we need to start with sender's or receiver's
list of owned names and match each of the names to a selection
of rules.

Change-Id: I2854f6cc51b26ff04a9984778f899ba9656ba290