bus/policy: direct checking of policy, without copies
Instead of building a copy of policy for each client,
we can check it using the main copy. Client copies
are rebuilt when the main copy is changed, anyway.

This removes copying of the main policy to client copies.
Instead, copy data that was used to select rules,
and use it for selection of rules from the main copy.

Change-Id: I42926c107aae0be1a1247a61f3558122b07f9914