cmake: added "AUTO" option when generating doxygen based documentation
This option, which is set by default, builds doxygen based documentation
if the required tool was found on configuring.

As part of this change, the name of the configuration parameter for
generating the documentation also had to be changed to ENABLE_DOXYGEN_DOCS,
which is more in line with the corresponding Autotools option.