cmake: Find docbook-xsl stylesheets using URI
Looking at random paths in CMAKE_INSTALL_DATAROOTDIR will only work
when the XSL stylesheets are installed to the same prefix as dbus,
which is not the case on Nix.

Let’s use libxml’s ability to look the stylesheet in XML catalog.
It will check the system’s global catalog (/etc/xml/catalog),
when not cross-compiling, or in one of the catalogs listed
in `XML_CATALOG_FILES` environment variable, when set.