Add support to send verbose messages to lttng
The support is enabled by adding -DENABLE_VERBOSE_LTTNG=1 to the cmake
configure line. After building with this configuration, dbus verbose
messages are passed to the lttng tracing system. A separate test program
called `test-lttng` is included to verify that support is working.

To add lttng tracepoints to a library or application, the cmake macro
add_lttng_tracepoint() is available. It is a wrapper around the lttng
command line tool `lttng-gen-tp` that generates C and header files from
a lttng-tracepoint file, e.g. `lttng-xxx.tp`, which are added to the
libraries and applications to be built. See `tests/CMakeLists.txt` for an
example of how to use it.

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>