subprojects: Update fallback expat subproject
Updated with:

    meson wrap update

expat 2.5.0 contains security fixes. In practice these don't affect
dbus, because we only use expat to parse trusted message bus
configuration, but we might as well use the newest available version.

For now I'm holding back GLib to the version we're currently using,
to avoid a dependency on a newer Meson and a build failure on MSVC
(see dbus!426).

Signed-off-by: Simon McVittie <smcv@collabora.com>