Add a limit for pending messages to a single service
Allows to define the maximum number of pending messages for a single
destination service. It extends the existing functionality of checking
the number of pending messages from the sender in the message queue. By
setting the maximum number of pending messages for a single service can
prevent reaching the limit of messages sent by service in case when one
of the destination services doesn't respond within the timeout limit,
thus preventing the next messages from the sender from being skipped.

Signed-off-by: Marek Maslanka <mmaslanka@chromium.org>